Feat: 엔트리 뷰 컴포넌트 tailwind로 변경
